[스레드뷰] 서버 연결 및 부가 기능 추가

+++ android/app/src/main/java/com/created/team201/presentation/home/adapter/HomeAdapter.kt
@@ -7,7 +7,7 @@ import com.created.domain.model.UserStudy
 
 
 class HomeAdapter(
-    private val onClick: (studyId: Int) -> Unit
+    private val onClick: (studyId: Long) -> Unit
 ) : ListAdapter<UserStudy, HomeViewHolder>(diffCallBack) {
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): HomeViewHolder {
