+++ android/domain/src/main/java/com/created/domain/model/UserStudy.kt
@@ -1,7 +1,7 @@
 package com.created.domain.model
 
 data class UserStudy(
-    val studyId: Int,
+    val studyId: Long,
     val studyName: String,
     val mustDo: String,
     val leftDays: Int,
